--- Backend/WebApp/config.py
@@ -1,1 +1,0 @@
-#Will be used for any configs needed
