--- Backend/WebApp/forms.py
@@ -1,1 +1,0 @@
-#Used for forms in the future if needed
