--- Backend/WebApp/models.py
@@ -1,1 +1,0 @@
-#Used for database models from PostgreSQL and others
