--- Backend/WebApp/routes.py
@@ -1,13 +1,0 @@
-from WebApp import app
-from flask import send_from_directory
-import os
-from PredictionModel import stock_api as API
-
-# Serve React App
-@app.route('/', defaults={'path': ''})
-@app.route('/<path:path>')
-def serve(path):
-    if path != "" and os.path.exists(os.path.join(app.static_folder, path)):
-        return send_from_directory(app.static_folder, path)
-    else:
-        return send_from_directory(app.static_folder, 'index.html')
