+++ Backend/WebApp/test.py
@@ -1,0 +1,16 @@
+import os
+
+# Define the relative path
+relative_path = '../../frontend/profitpulsex'
+
+# Get the absolute path
+absolute_path = os.path.abspath(relative_path)
+
+# Print the absolute path
+print(f"Absolute path: {absolute_path}")
+
+# Check if the path exists
+if os.path.exists(absolute_path):
+    print("The path exists.")
+else:
+    print("The path does not exist.")
